% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.kcsap}
\alias{process.kcsap}
\title{Process KCSAP data.}
\usage{
process.kcsap(aws_dir, kcsap_dir, adt_dir)
}
\arguments{
\item{aws_dir}{full path to the directory ADT.}

\item{kcsap_dir}{full path to the directory containing the KCSAP data.}

\item{adt_dir}{full path to the directory containing the AWS_DATA folder on ADT server.}
}
\description{
Get the data from KCSAP, parse and convert into ADT format.
}
