% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.tahmo}
\alias{process.tahmo}
\title{Process TAHMO data from API.}
\usage{
process.tahmo(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Get the data from TAHMO API, parse and convert into ADT format.
}
